refactor:  getUserById와 getEventById를 전용 Provider 클래스로 이동
- 서비스 레이어 간 순환 참조를 방지하고, 공통 조회 로직을 한 곳에 모으기 위해 Provider 클래스로 분리
- LikeService, VisitService 등에서 재사용 가능하도록 개선
