+++ src/main/java/com/moonbaar/domain/event/service/EventService.java
@@ -5,8 +5,7 @@ import com.moonbaar.domain.event.dto.EventListResponse;
 import com.moonbaar.domain.event.dto.EventSearchRequest;
 import com.moonbaar.domain.event.dto.EventSummaryResponse;
 import com.moonbaar.domain.event.entity.CulturalEvent;
-import com.moonbaar.domain.event.exeption.EventNotFoundException;
-import com.moonbaar.domain.event.repository.CulturalEventRepository;
+import com.moonbaar.domain.event.repository.EventRepository;
 import com.moonbaar.domain.event.repository.EventSpecifications;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -21,7 +20,8 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class EventService {
 
-    private final CulturalEventRepository eventRepository;
+    private final EventRepository eventRepository;
+    private final EventProvider eventProvider;
 
     public EventListResponse searchEvents(EventSearchRequest request) {
         Specification<CulturalEvent> spec = createSearchSpecification(request);
@@ -54,12 +54,7 @@ public class EventService {
     }
 
     public EventDetailResponse getEventDetail(Long eventId) {
-        CulturalEvent event = getEventById(eventId);
+        CulturalEvent event = eventProvider.getEventById(eventId);
         return EventDetailResponse.from(event);
-    }
-
-    public CulturalEvent getEventById(Long eventId) {
-        return eventRepository.findById(eventId)
-                .orElseThrow(EventNotFoundException::new);
     }
 }
