+++ src/main/java/com/moonbaar/domain/like/service/LikeService.java
@@ -1,7 +1,7 @@
 package com.moonbaar.domain.like.service;
 
 import com.moonbaar.domain.event.entity.CulturalEvent;
-import com.moonbaar.domain.event.service.EventService;
+import com.moonbaar.domain.event.service.EventProvider;
 import com.moonbaar.domain.like.dto.LikeResponse;
 import com.moonbaar.domain.like.dto.LikedEventListRequest;
 import com.moonbaar.domain.like.dto.LikedEventListResponse;
@@ -10,7 +10,7 @@ import com.moonbaar.domain.like.exception.AlreadyLikedEventException;
 import com.moonbaar.domain.like.exception.LikeNotFoundException;
 import com.moonbaar.domain.like.repository.LikedEventRepository;
 import com.moonbaar.domain.user.entity.User;
-import com.moonbaar.domain.user.service.UserService;
+import com.moonbaar.domain.user.service.UserProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -21,14 +21,14 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class LikeService {
 
-    private final UserService userService;
-    private final EventService eventService;
+    private final UserProvider userProvider;
+    private final EventProvider eventProvider;
     private final LikedEventRepository likedEventRepository;
 
     @Transactional
     public LikeResponse likeEvent(Long userId, Long eventId) {
-        User user = userService.getUserById(userId);
-        CulturalEvent event = eventService.getEventById(eventId);
+        User user = userProvider.getUserById(userId);
+        CulturalEvent event = eventProvider.getEventById(eventId);
 
         checkNotAlreadyLiked(user, event);
 
@@ -43,8 +43,8 @@ public class LikeService {
 
     @Transactional
     public LikeResponse unlikeEvent(Long userId, Long eventId) {
-        User user = userService.getUserById(userId);
-        CulturalEvent event = eventService.getEventById(eventId);
+        User user = userProvider.getUserById(userId);
+        CulturalEvent event = eventProvider.getEventById(eventId);
 
         LikedEvent likedEvent = findLikedEvent(user, event);
         likedEventRepository.delete(likedEvent);
@@ -65,7 +65,7 @@ public class LikeService {
 
     @Transactional(readOnly = true)
     public LikedEventListResponse getLikedEvents(Long userId, LikedEventListRequest request) {
-        User user = userService.getUserById(userId);
+        User user = userProvider.getUserById(userId);
         Pageable pageable = request.toPageable();
 
         Page<LikedEvent> likedEventsPage = likedEventRepository.findByUser(user, pageable);
