+++ src/main/java/com/moonbaar/domain/user/service/UserProvider.java
@@ -1,0 +1,19 @@
+package com.moonbaar.domain.user.service;
+
+import com.moonbaar.domain.user.entity.User;
+import com.moonbaar.domain.user.exception.UserNotFoundException;
+import com.moonbaar.domain.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserProvider {
+
+    private final UserRepository userRepository;
+
+    public User getUserById(Long userId) {
+        return userRepository.findById(userId)
+                .orElseThrow(UserNotFoundException::new);
+    }
+}
